Fix issue 3993
